Create simple-paragraph
This is the paragraph level parsing in PHP for a very simple paragraph. It doesn't yet contain introspection or handle more complex objects like links or citations, and relies on object keys in JSON being identical to HTML. It is really intended as a very basic introduction to the theory and structure of paragraphs, and to also demonstrate a commitment to keep document, paragraph and character styles separated from one another so as to avoid confusion between the functions.